refactor: SearchActivity가 가지지 않도록 수정
