+++ android/app/src/main/java/com/on/turip/ui/search/keywordresult/SearchActivity.kt
@@ -1,161 +1,93 @@
 package com.on.turip.ui.search.keywordresult
 
-import android.content.Context
-import android.content.Intent
-import android.graphics.Rect
-import android.os.Bundle
 import android.text.Editable
-import android.view.MenuItem
-import android.view.MotionEvent
-import android.view.View
-import android.view.inputmethod.EditorInfo
-import android.view.inputmethod.InputMethodManager
-import android.widget.EditText
-import androidx.activity.viewModels
-import androidx.core.content.ContextCompat
-import androidx.core.widget.addTextChangedListener
-import com.on.turip.R
-import com.on.turip.databinding.ActivitySearchBinding
-import com.on.turip.ui.common.base.BaseActivity
-import com.on.turip.ui.trip.detail.TripDetailActivity
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.viewmodel.initializer
+import androidx.lifecycle.viewmodel.viewModelFactory
+import com.on.turip.di.RepositoryModule
+import com.on.turip.domain.content.PagedContentsResult
+import com.on.turip.domain.content.repository.ContentRepository
+import com.on.turip.domain.content.video.VideoInformation
+import com.on.turip.ui.common.mapper.toUiModel
+import com.on.turip.ui.search.model.VideoInformationModel
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.async
+import kotlinx.coroutines.launch
+import timber.log.Timber
 
-class SearchActivity : BaseActivity<ActivitySearchBinding>() {
-    private val viewModel: SearchViewModel by viewModels {
-        SearchViewModel.provideFactory()
+class SearchViewModel(
+    private val contentRepository: ContentRepository,
+) : ViewModel() {
+    private val _searchingWord: MutableLiveData<String> = MutableLiveData()
+    val searchingWord: LiveData<String> = _searchingWord
+
+    private val _videoInformation: MutableLiveData<List<VideoInformationModel>> = MutableLiveData()
+    val videoInformation: LiveData<List<VideoInformationModel>> = _videoInformation
+
+    private val _searchResultCount: MutableLiveData<Int> = MutableLiveData()
+    val searchResultCount: LiveData<Int> = _searchResultCount
+
+    private val _loading: MutableLiveData<Boolean> = MutableLiveData()
+    val loading: LiveData<Boolean> = _loading
+
+    fun updateSearchingWord(newWords: Editable?) {
+        _searchingWord.value = newWords.toString()
     }
 
-    override val binding: ActivitySearchBinding by lazy {
-        ActivitySearchBinding.inflate(layoutInflater)
-    }
-
-    private val searchAdapter: SearchAdapter =
-        SearchAdapter { contentId: Long?, creatorId: Long? ->
-            val intent: Intent =
-                TripDetailActivity.newIntent(
-                    context = this,
-                    contentId = contentId ?: 0,
-                    creatorId = creatorId ?: 0,
-                )
-            startActivity(intent)
-        }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setupToolbar()
-        setupListeners()
-        setupObserves()
-        binding.etSearchResult.requestFocus()
-        binding.rvSearchResult.adapter = searchAdapter
-        binding.rvSearchResult.itemAnimator = null
-    }
-
-    private fun setupToolbar() {
-        setSupportActionBar(binding.tbSearchResult)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-
-        binding.tbSearchResult.navigationIcon?.setTint(
-            ContextCompat.getColor(
-                this,
-                R.color.gray_300_5b5b5b,
-            ),
-        )
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == android.R.id.home) {
-            finish()
-            return true
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    private fun setupListeners() {
-        binding.etSearchResult.addTextChangedListener { editable: Editable? ->
-            viewModel.updateSearchingWord(editable)
-        }
-        binding.etSearchResult.setOnEditorActionListener { _, actionId, _ ->
-            if (actionId == EditorInfo.IME_ACTION_SEARCH) {
-                viewModel.updateByKeyword()
-                val inputMethodManager =
-                    getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
-                inputMethodManager.hideSoftInputFromWindow(binding.etSearchResult.windowToken, 0)
-                true
-            } else {
-                false
-            }
-        }
-
-        binding.ivSearchResultClear.setOnClickListener {
-            binding.etSearchResult.text.clear()
-        }
-    }
-
-    private fun setupObserves() {
-        var latestSearchResultCount: Int = 0
-        var latestLoading: Boolean = false
-
-        viewModel.searchResultCount.observe(this) { count ->
-            latestSearchResultCount = count
-            binding.tvSearchResultCount.text =
-                getString(R.string.search_result_exist_result, count)
-            updateUiState(latestLoading, count)
-        }
-
-        viewModel.loading.observe(this) { loading ->
-            latestLoading = loading
-            updateUiState(loading, latestSearchResultCount)
-        }
-
-        viewModel.videoInformation.observe(this) { videoInformationModels ->
-            searchAdapter.submitList(videoInformationModels)
-        }
-
-        viewModel.searchingWord.observe(this) { words ->
-            binding.ivSearchResultClear.visibility =
-                if (words.isNotBlank()) View.VISIBLE else View.GONE
-        }
-    }
-
-    private fun updateUiState(
-        loading: Boolean,
-        searchResultCount: Int,
-    ) {
-        if (loading) {
-            binding.pbSearchResult.visibility = View.VISIBLE
-            binding.tvSearchResultCount.visibility = View.GONE
-            binding.rvSearchResult.visibility = View.GONE
-            binding.groupSearchResultEmpty.visibility = View.GONE
-        } else {
-            binding.pbSearchResult.visibility = View.GONE
-            if (searchResultCount == 0) {
-                binding.groupSearchResultEmpty.visibility = View.VISIBLE
-                binding.tvSearchResultCount.visibility = View.GONE
-                binding.rvSearchResult.visibility = View.GONE
-            } else {
-                binding.groupSearchResultEmpty.visibility = View.GONE
-                binding.tvSearchResultCount.visibility = View.VISIBLE
-                binding.rvSearchResult.visibility = View.VISIBLE
-            }
-        }
-    }
-
-    override fun dispatchTouchEvent(ev: MotionEvent): Boolean {
-        if (ev.action == MotionEvent.ACTION_DOWN) {
-            val v = currentFocus
-            if (v is EditText) {
-                val outRect = Rect()
-                v.getGlobalVisibleRect(outRect)
-                if (!outRect.contains(ev.rawX.toInt(), ev.rawY.toInt())) {
-                    v.clearFocus()
-                    val imm = getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
-                    imm.hideSoftInputFromWindow(v.windowToken, 0)
+    fun updateByKeyword() {
+        viewModelScope.launch {
+            val searchResultCountResult =
+                async {
+                    _loading.value = true
+                    contentRepository.loadContentsSizeByKeyword(
+                        searchingWord.value.toString(),
+                    )
                 }
-            }
+
+            val pagedContentsResult: Deferred<Result<PagedContentsResult>> =
+                async {
+                    contentRepository.loadContentsByKeyword(
+                        keyword = searchingWord.value.toString(),
+                        size = 100,
+                        lastId = 0L,
+                    )
+                }
+
+            pagedContentsResult
+                .await()
+                .onSuccess { result: PagedContentsResult ->
+                    val videoModels: List<VideoInformationModel> =
+                        result.videos.map { videoInformation: VideoInformation -> videoInformation.toUiModel() }
+                    _videoInformation.value = videoModels
+                }.onFailure {
+                    _loading.value = false
+                    Timber.e("${it.message}")
+                }
+
+            searchResultCountResult
+                .await()
+                .onSuccess { result: Int ->
+                    _loading.value = false
+                    _searchResultCount.value = result
+                }.onFailure {
+                    _loading.value = false
+                    Timber.e("${it.message}")
+                }
         }
-        return super.dispatchTouchEvent(ev)
     }
 
     companion object {
-        fun newIntent(context: Context): Intent = Intent(context, SearchActivity::class.java)
+        fun provideFactory(contentRepository: ContentRepository = RepositoryModule.contentRepository): ViewModelProvider.Factory =
+            viewModelFactory {
+                initializer {
+                    SearchViewModel(
+                        contentRepository,
+                    )
+                }
+            }
     }
 }
